[grug] Split the ragged gate by what crosses the wire
The gate's callable is pickled by reference, so the GB200 worker imports
whatever module defines it. Putting that in a pytest module made the worker
import a test file; `marin.testing.inference` already exists because the
snowball cluster test hit the same thing and kept its remote callable out of
the test. Follow that: `marin.testing.moe.ragged_ep_gate` holds everything that
ships, and the test under tests/cluster/grug only submits it.

Two things fall out. The unit tests for the drop oracle now import a packaged
module instead of reaching into a cluster test. And the XLA flags move to
`levanter.grug._moe.ep_ragged_all_to_all`, which is where the backend they
select lives -- a marin module importing them from experiments/ would have run
the dependency direction backwards.

The gate takes its target cluster from the caller, since naming a peer is the
submitter's business.
