+++ lib/marin/src/marin/testing/moe/__init__.py
@@ -1,0 +1,4 @@
+# Copyright The Marin Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Reusable mixture-of-experts validation support."""
